+++ docs/changelog.md
@@ -4,6 +4,7 @@
 - support `now` for duration specificatoins
 - `infr expiry` is now relative to the current time
 - add `.keeper actions` command
+- support command loading and unloading at runtime through the `sudo` cog
 
 ## v0.11.1
 - removed default `.clean` invocation