+++ docs/changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+- added the `.lastjoins` command
 - support `now` for duration specificatoins
 - `infr expiry` is now relative to the current time
 - add `.keeper actions` command