+++ docs/changelog.md
@@ -6,6 +6,8 @@
 - support `now` for duration specificatoins
 - `infr expiry` is now relative to the current time
 - add `.keeper actions` command
+
+# v0.11.2
 - support command loading and unloading at runtime through the `sudo` cog
 - prevent `role` alias to `roleinfo` from shadowing `role [subcommand]`
 - display more specific error when `.assign` cannot find command author in guild